requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2023.11.17
colorama==0.4.6
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.47.2
fpdf2==2.7.7
frozenlist==1.4.1
gino==1.0.1
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.2
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
magic-filter==1.0.12
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
pure-eval==0.2.2
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==306
pyzmq==25.1.2
six==1.16.0
SQLAlchemy==1.3.24
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
wcwidth==0.2.13
yarl==1.9.4
zipp==3.17.0
